refactor: align site root util with render-page refactor
