+++ lib/copy-asset.js
@@ -1,7 +1,8 @@
-import { basename, dirname, join, relative, fromFileUrl } from "@std/path";
+import { basename, dirname, fromFileUrl, join, relative } from "@std/path";
 import { SRC_ASSET_EXTENSIONS } from "./extension-whitelist.js";
 import { hashAssetName } from "./hash-asset.js";
 import { getEmoji } from "./emoji.js";
+import { findSiteRoot } from "./load-config.js";
 
 /**
  * Copy an asset from the source tree to its distant output directory.
@@ -31,7 +32,9 @@ export async function copyAsset(path) {
     await Deno.stat(srcPath);
   } catch (err) {
     if (err instanceof Deno.errors.NotFound && ext === ".css") {
-      const fallback = fromFileUrl(new URL(`../core/css/${rel}`, import.meta.url));
+      const fallback = fromFileUrl(
+        new URL(`../core/css/${rel}`, import.meta.url),
+      );
       try {
         await Deno.stat(fallback);
         srcPath = fallback;
@@ -133,28 +136,5 @@ export async function removeAsset(path) {
     await Deno.remove(outPath);
   } catch (err) {
     if (!(err instanceof Deno.errors.NotFound)) throw err;
-  }
-}
-
-/**
- * Walk up the directory tree looking for a `config.json` file.
- *
- * @param {string} filePath Path of the file whose site root is required.
- * @returns {Promise<string>} Directory containing the `config.json` file.
- */
-async function findSiteRoot(filePath) {
-  let dir = dirname(filePath);
-  while (true) {
-    try {
-      const stat = await Deno.stat(join(dir, "config.json"));
-      if (stat.isFile) return dir;
-    } catch (err) {
-      if (!(err instanceof Deno.errors.NotFound)) throw err;
-    }
-    const parent = dirname(dir);
-    if (parent === dir) {
-      throw new Error(`config.json not found for ${filePath}`);
-    }
-    dir = parent;
   }
 }
