+++ lib/render-page.js
@@ -1,13 +1,13 @@
-import { dirname, join, relative, toFileUrl, fromFileUrl } from "@std/path";
-import { hashAssetName } from "./hash-asset.js";
-import { copyAsset } from "./copy-asset.js";
-import { DOMParser } from "@b-fuze/deno-dom";
+import { join, relative } from "@std/path";
 import { parsePage } from "./parse-page.js";
-import { LinksManager } from "./links.js";
-import { applyTemplates } from "./apply-templates.js";
-import { inlineSvg } from "./inline-svg.js";
-import { getEmoji } from "./emoji.js";
 import { markdownToHTML } from "./markdown.js";
+import { getEmoji } from "./emoji.js";
+import { loadConfig, findSiteRoot } from "./load-config.js";
+import { processCss } from "./process-css.js";
+import { processModules } from "./process-modules.js";
+import { manageLinks } from "./manage-links.js";
+import { buildDocument } from "./build-document.js";
+import { writeOutput, toOutRel } from "./write-output.js";
 
 /**
  * @typedef {object} RenderResult
@@ -43,150 +43,31 @@ export async function renderPage(path, root = new URL("..", import.meta.url)) {
       }
     }
 
-    const siteDir = await findSiteRoot(path);
-    const configPath = join(siteDir, "config.json");
-    const configText = await Deno.readTextFile(configPath);
-    const config = JSON.parse(configText);
+    const { siteDir, config } = await loadConfig(path);
     const distant = String(config.distantDirectory);
     const pretty = Boolean(config.prettyUrls);
     const hashAssets = Boolean(config.hashAssets);
 
-    const cssUsed = [];
-    const cssFiles = page.frontMatter.css ?? [];
-    page.frontMatter.css = await Promise.all(
-      cssFiles.map(async (href) => {
-        const relHref = href.startsWith("/") ? href.slice(1) : href;
-        const abs = join(siteDir, relHref);
-        try {
-          const real = await Deno.realPath(abs);
-          cssUsed.push(real);
-          if (!hashAssets) return href;
-          const hashed = await hashAssetName(real);
-          const dirRel = dirname(relHref);
-          const relPath = (dirRel === "." ? hashed : join(dirRel, hashed))
-            .replace(/\\/g, "/");
-          return href.startsWith("/") ? "/" + relPath : relPath;
-        } catch {
-          const corePath = fromFileUrl(
-            new URL(`../core/css/${relHref}`, import.meta.url),
-          );
-          try {
-            await Deno.stat(corePath);
-            cssUsed.push(abs);
-            if (hashAssets) {
-              const hashed = await hashAssetName(corePath);
-              const dirRel = dirname(relHref);
-              const relPath = (dirRel === "." ? hashed : join(dirRel, hashed))
-                .replace(/\\/g, "/");
-              await copyAsset(abs);
-              return href.startsWith("/") ? "/" + relPath : relPath;
-            }
-            await copyAsset(abs);
-            return href;
-          } catch {
-            console.log(`${getEmoji("error")} CSS missing -- ${relHref}`);
-            cssUsed.push(abs);
-            return href;
-          }
-        }
-      }),
-    );
-    const modulesUsed = [];
-    const moduleFiles = page.scripts.modules ?? [];
-    page.scripts.modules = await Promise.all(
-      moduleFiles.map(async (src) => {
-        const relSrc = src.startsWith("/") ? src.slice(1) : src;
-        const abs = join(siteDir, relSrc);
-        let real;
-        try {
-          real = await Deno.realPath(abs);
-        } catch {
-          real = abs;
-        }
-        modulesUsed.push(real);
-        if (!hashAssets) return src;
-        const hashed = await hashAssetName(abs);
-        const dirRel = dirname(relSrc);
-        const relPath = (dirRel === "." ? hashed : join(dirRel, hashed))
-          .replace(/\\/g, "/");
-        return src.startsWith("/") ? "/" + relPath : relPath;
-      }),
-    );
+    const cssUsed = await processCss(page, siteDir, hashAssets);
+    const modulesUsed = await processModules(page, siteDir, hashAssets);
 
-    const linksManager = new LinksManager(join(siteDir, "links.json"));
-    await linksManager.load();
-    const rel = relative(siteDir, path).replace(/\\/g, "/");
-    const href = toHref(rel, pretty);
-    const linksUsed = page.frontMatter.links !== undefined;
-    // Always merge so links are pruned when omitted from the page.
-    const linksChanged = linksManager.merge(href, page.links ?? {});
-    if (linksChanged) await linksManager.save();
-
-    const parser = new DOMParser();
-    // Build a minimal document. Templates are responsible for generating
-    // the final <head> and <body> structure.
-    const doc = parser.parseFromString("<html></html>", "text/html");
-    if (!doc) throw new Error(`${path}: invalid HTML`);
-
-    const templatesUsed = await applyTemplates(
-      doc,
+    const { linksManager, linksUsed, linksChanged } = await manageLinks(
       page,
-      linksManager.data,
+      siteDir,
+      path,
+      pretty,
+    );
+
+    const { doc, templatesUsed, scriptsUsed, svgsUsed } = await buildDocument(
+      page,
+      linksManager,
       config,
       root,
+      siteDir,
+      path,
     );
 
-    const body = doc.querySelector("body");
-    if (!body) throw new Error(`${path}: missing <body> element`);
-
-    for (const src of page.scripts.modules ?? []) {
-      const script = doc.createElement("script");
-      script.setAttribute("type", "module");
-      script.setAttribute("src", src);
-      body.appendChild(script);
-    }
-
-    const scriptsUsed = [];
-    for (const file of page.scripts.inline ?? []) {
-      const scriptPath = join(siteDir, file);
-      let realPath;
-      try {
-        realPath = await Deno.realPath(scriptPath);
-      } catch {
-        realPath = scriptPath;
-      }
-      let content;
-      try {
-        content = await Deno.readTextFile(realPath);
-      } catch (err) {
-        if (err instanceof Error) err.message = `${realPath}: ${err.message}`;
-        throw err;
-      }
-      scriptsUsed.push(realPath);
-      const script = doc.createElement("script");
-      script.textContent = content;
-      body.appendChild(script);
-    }
-
-    const svgsUsed = await inlineSvg(doc, toFileUrl(siteDir + "/"));
-
-    const outRel = toOutRel(rel, pretty);
-    const outPath = join(distant, outRel);
-    await Deno.mkdir(dirname(outPath), { recursive: true });
-    const htmlOut = "<!DOCTYPE html>\n" + doc.documentElement.outerHTML;
-    await Deno.writeTextFile(outPath, htmlOut);
-
-    const fmtCmd = new Deno.Command(Deno.execPath(), {
-      args: ["fmt", outPath],
-      stdout: "null",
-      stderr: "piped",
-    });
-    const { code, stderr } = await fmtCmd.output();
-    if (code !== 0) {
-      throw new Error(
-        `${outPath}: ${new TextDecoder().decode(stderr).trim()}`,
-      );
-    }
+    await writeOutput(doc, path, siteDir, distant, pretty);
 
     return {
       pagePath: path,
@@ -228,61 +109,5 @@ export async function removePage(path) {
     await Deno.remove(outPath);
   } catch (err) {
     if (!(err instanceof Deno.errors.NotFound)) throw err;
-  }
-}
-
-/**
- * Build the href used in `links.json` for a page.
- *
- * @param {string} rel Page path relative to the site root.
- * @param {boolean} pretty Whether `.html` extensions should be omitted.
- * @returns {string} Href beginning with a leading slash.
- */
-function toHref(rel, pretty) {
-  const normalized = rel.replace(/\\/g, "/").replace(/\.(md|html?)$/i, ".html");
-  if (!pretty) return "/" + normalized;
-  return "/" +
-    normalized.replace(/index\.html$/i, "").replace(/\.html$/i, "/");
-}
-
-/**
- * Determine output file path relative to the distant directory.
- *
- * @param {string} rel Page path relative to the site root.
- * @param {boolean} pretty When true, preserve the original file structure
- * without creating nested `index.html` directories.
- * @returns {string} Relative output path within the distant directory.
- */
-function toOutRel(rel, pretty) {
-  const normalized = rel.replace(/\\/g, "/");
-  if (pretty) {
-    // Ensure the file keeps a `.html` extension without nesting it under an
-    // additional directory. Previously this would transform `foo.html` into
-    // `foo/index.html` which broke references to the generated file.
-    return normalized.replace(/\.(md|html?)$/i, ".html");
-  }
-  return normalized.replace(/\.(md|html?)$/i, "") + ".html";
-}
-
-/**
- * Walk up the directory hierarchy to find the site root containing `config.json`.
- *
- * @param {string} filePath Starting path used to search for the site root.
- * @returns {Promise<string>} Directory path containing `config.json`.
- */
-async function findSiteRoot(filePath) {
-  let dir = dirname(filePath);
-  while (true) {
-    try {
-      const stat = await Deno.stat(join(dir, "config.json"));
-      if (stat.isFile) return dir;
-    } catch (err) {
-      if (!(err instanceof Deno.errors.NotFound)) throw err;
-    }
-    const parent = dirname(dir);
-    if (parent === dir) {
-      throw new Error(`config.json not found for ${filePath}`);
-    }
-    dir = parent;
   }
 }
